+++ gmsrc/src/gm/gmMachineLib.cpp
@@ -742,7 +742,7 @@ static int GM_CDECL gmTableDuplicate(gmThread * a_thread)
 
 void gmConcat(gmMachine * a_machine, char * &a_dst, int &a_len, int &a_size, const char * a_src, int a_growBy = 32)
 {
-  int len = strlen(a_src);
+  int len = (int)strlen(a_src);
 
   if((a_len + len + 1) >= a_size)
   {
